Create encounterUtils and move all the encounter methods from utils to it
