+++ server/config/utils.js
@@ -8,28 +8,4 @@ var bookshelf = require('bookshelf');
 
 module.exports = {
 
-  createEncounter: function() {
-    // TODO add description to schema.sql encounter table
-    Encounters.create({
-      title: req.body.title,
-      location: req.body.location,
-      encountertime: req.body.encountertime,
-      posttime: new Date()
-    })
-    .then(function(encounter) {
-      res.status(200).send(encounter);
-    })
-    .catch(function(error) {
-      res.status(500).send(error.message);
-    });
-  },
-
-  showAllEncounters: function() {
-
-  },
-
-  recentActivity: function() {
-
-  }
-
 };
